requirements.txt
google-generativeai
opencv-python
numpy
mss
python-dotenv 
pillow
pyautogui
pytesseract
